\name{coleR-package}
\alias{coleR-package}
\alias{coleR}
\docType{package}
\title{
\packageTitle{coleR}
}
\description{
\packageDescription{coleR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{coleR}
\packageIndices{coleR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{coleR}

Maintainer: \packageMaintainer{coleR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
