% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\alias{fuzzy_check}
\title{Link data.tables by fuzzy string matching}
\usage{
fuzzy_match(a, b, method = "jw", cutoff = 0.5, ...)

fuzzy_check(a, b, method = "jw", ...)
}
\arguments{
\item{a}{a source vector of strings}

\item{b}{a target vector}

\item{method}{method for \code{\link[stringdist]{stringdistmatrix}}}

\item{cutoff}{numeric indicating the maximum distance threshold for a match (\code{fuzzy_match} only).
String distances equal to or below the cutoff are counted as matches.}

\item{...}{further arguments for \code{\link[stringdist]{stringdistmatrix}}}
}
\value{
For \code{fuzzy_match}, a vector of nearest string matches.
For strings with multiple closest matches, only the first is returned.

For \code{fuzzy_check}, a data.frame containing the source strings,
their closest matches, and the string distance for each match.
}
\description{
Finds the closest string match.
The default method computes Jaro-Winkler string distances using the \code{stringdist} package.
For strings with multiple closest matches, only the first is reported.
}
\examples{
library(data.table)

set.seed(575)
fruit <- sample(stringr::fruit, 30)

DTA <- data.table(block1 = sample(LETTERS[1:4], 20, TRUE),
                  block2 = sample(LETTERS[1:4], 20, TRUE),
                  fruit   = sample(fruit, 20))

DTB <- data.table(block1 = sample(LETTERS[1:4], 20, TRUE),
                  block2 = sample(LETTERS[1:4], 20, TRUE),
                  fruit   = sample(fruit, 20))

fuzzy_check(DTA$fruit, DTB$fruit)
fuzzy_match(DTA$fruit, DTB$fruit)

setkey(DTB, block1, block2)

DTA[ , fuzzy_match(fruit, b = DTB[.BY, fruit]),
     by = .(block1, block2)]

}
\seealso{
\code{\link[stringdist]{stringdistmatrix}}
}
