% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grow.matrix.R
\name{grow.matrix}
\alias{grow.matrix}
\title{Grow Matrix}
\usage{
grow.matrix(x, data = NA, nrow = 1, ncol = nrow)
}
\arguments{
\item{x}{Matrix.}

\item{data}{Value to insert into new elements.}

\item{nrow}{Integer. Number of rows to add.}

\item{ncol}{Integer. Number of columns to add.}
}
\description{
Extend the rows and columns of a matrix by a value.
}
\examples{
mat <- matrix(seq(16), nrow=4)
grow.matrix(mat, nrow=2)
grow.matrix(mat, 0, nrow=4, ncol=2)
grow.matrix(mat, -1, nrow=0, ncol=2)
}

