% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listtodf.R
\name{listtodf}
\alias{listtodf}
\title{List to data.frame}
\usage{
listtodf(x)
}
\arguments{
\item{x}{list containing data.frames}
}
\value{
data.frame
}
\description{
Convert a list containing similar data.frames into a single
data.frame object.
}
\examples{
df <- data.frame(id=1:10, val=rnorm(10))
x <- list(df, df[sample(nrow(df)),], df[sample(nrow(df), 4),])
listtodf(x)
}

