% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcheck.R
\name{mcheck}
\alias{mcheck}
\alias{merge_check}
\title{Check match rates}
\usage{
mcheck(x, y, incomparables = FALSE)
}
\arguments{
\item{x}{a character vector}

\item{y}{a character vector}

\item{incomparables}{a vector of incomparable values, where FALSE means no values.
See \code{\link[base]{unique}}}
}
\value{
A numeric matrix of match statistics.
}
\description{
Check match rates between two character vectors
}
\examples{
A <- sample(state.abb, 50, replace = TRUE)
B <- sample(state.abb, 100, replace = TRUE)
B[sample(1:length(B), 10)] <- NA_character_

mcheck(A, B)

}
\seealso{

}
