% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prettyprint.R
\name{prettyprint}
\alias{prettyprint}
\title{Pretty Print}
\usage{
prettyprint(x, labels)
}
\arguments{
\item{x}{A data.frame or matrix. Object to print.}

\item{labels}{A character vector. Column names.}
}
\description{
Print objects with a SQL-like format.
}
\examples{
df <- data.frame(id=1:10, vals=rnorm(10))
prettyprint(df)
prettyprint(df, c('ID','RANDOM VALUES'))
}

