% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rda.R
\name{read_rda}
\alias{read_rda}
\title{Read an .rda file and return its contents}
\usage{
read_rda(file)
}
\arguments{
\item{file}{a string with the file name and path.}
}
\value{
An object or list of objects stored in \code{file}.
}
\description{
Reads the contents of an .RData file into a separate environment
(to avoid overwriting existing variables) and returns the contents for assignment
to your own variable names.
}
