% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{safely}
\alias{object_size}
\alias{compare}
\alias{map}
\alias{str_extract}
\alias{str_split}
\alias{str_split_fixed}
\alias{skim}
\alias{tabyl}
\alias{tidy}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{broom}{\code{\link[broom]{tidy}}}

  \item{janitor}{\code{\link[janitor]{tabyl}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{pryr}{\code{\link[pryr]{object_size}}}

  \item{purrr}{\code{\link[purrr]{safely}}, \code{\link[purrr]{map}}}

  \item{skimr}{\code{\link[skimr]{skim}}}

  \item{stringr}{\code{\link[stringr]{str_extract}}, \code{\link[stringr]{str_split}}, \code{\link[stringr]{str_split_fixed}}}

  \item{waldo}{\code{\link[waldo]{compare}}}
}}

