% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_list.R
\name{save_list}
\alias{save_list}
\title{Write list elements to disk as separate objects}
\usage{
save_list(
  x,
  dir,
  ext = c(".fst", ".qs", ".rds", ".csv", ".pdf"),
  names = NULL,
  cores = 1,
  compress = 100,
  ...
)
}
\arguments{
\item{x}{a list of objects to save.}

\item{dir}{a directory in which to save the list elements.}

\item{ext}{a single extension for all files.}

\item{names}{a vector of file names without extensions. Defaults to the names of \code{x}.}

\item{cores}{number of cores to use in \code{\link{mcparallel}}. The default (1) runs in series. Note that several file-writing functions already implement multi-threading.}

\item{compress}{compression rate for \code{\link[fst]{write_fst}}.}

\item{...}{further arguments to pass to the file-writing function:
\code{\link{saveRDS}},
\code{\link[fst]{write_fst}}, \code{\link[qs]{qsave}},
\code{\link[data.table]{fwrite}},
or \code{\link[ggplot2]{ggsave}}.}
}
\description{
Write list elements to disk as separate objects
}
\examples{

tmp_dir <- tempdir()
df <- data.frame(char = LETTERS[1:10], num = 1:10)
ldf <- replicate(3, df, simplify = FALSE)

save_list(ldf, dir = tmp_dir, ext = ".csv")

list.files(tmp_dir)
unlink(tmp_dir, recursive = TRUE)

}
\seealso{
\code{\link{saveRDS}}, \code{\link[data.table]{fwrite}}, or \code{\link[ggplot2]{ggsave}}.
}
