% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{setContinue}
\alias{setContinue}
\title{Continue Prompt}
\usage{
setContinue(prompt = "    ")
}
\arguments{
\item{prompt}{Character. Defaults to four spaces.}
}
\description{
Set prompt for lines which continue over one line.
}

