% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{setMenuGraphics}
\alias{setMenuGraphics}
\title{Menu Graphics}
\usage{
setMenuGraphics(use = FALSE)
}
\arguments{
\item{use}{Logical. Defaults to FALSE.}
}
\description{
Set menu.graphics option.
}

