% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vsprintf.R
\name{vsprintf}
\alias{vsprintf}
\title{Vector sprintf}
\usage{
vsprintf(fmt, vals)
}
\arguments{
\item{fmt}{Character vector.  Argument to \code{sprintf}.}

\item{vals}{Vector of values to be displayed.}
}
\description{
Pass vector of values into \code{\link{sprintf}}.
}
\examples{
vsprintf("One year to a \%s is worth \%s years\%s", c('dog',7,'!'))
df <- data.frame(animal=c('dog','cat','fish'), years=c(7,5,13), ending=c('!','.','?'))
vsprintf("One year to a \%s is worth \%s years\%s", df)
}

