% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coleR-package.R
\docType{package}
\name{coleR-package}
\alias{coleR}
\alias{coleR-package}
\title{coleR}
\description{
Miscellaneous functions that do presumably helpful things.
}
\author{
Cole Beck

Maintainer: Cole Beck \email{cole.beck@gmail.com}
}

